Never cache dummy keys for addresses entered into reply-to
In a previous version a dummy reply-to key might be cached for an address and returned as the key to be used for a real recipient with the same address.

Example: A message is sent from lukele@gpgtools.org to lukele@gpgtools.org, cc’d to team@gpgtools.org and reply-to is set to team@gpgtools.org

If the reply-to field is compiled *before* the team@gpgtools.org address is entered into cc, the dummy key will have been cached for team@gpgtools.org and thus returned for the recipient team@gpgtools.org entered in cc.

As a result, since dummy keys are removed prior to sending, the message is only going to be encrypted to lukele@gpgtools.org, as team@gpgtools.org was pointing to a dummy key and thus has been removed.